LayoutTests/http/tests/security/contentSecurityPolicy/sandbox-in-http-header-control-expected.txt

ALERT: PASS: Iframe was not in a unique origin
